move darray shared. make shared a shared lib
