+++ CompilerSource/main.cpp
@@ -28,7 +28,7 @@ using namespace std;
 #define flushl '\n' << flush
 #define flushs flush
 
-#include "general/darray.h"
+#include "darray.h"
 
 #include "syntax/syncheck.h"
 #include "parser/parser.h"
