+++ CompilerSource/parser/object_storage.cpp
@@ -33,7 +33,7 @@
 
 using namespace std;
 
-#include "general/darray.h"
+#include "darray.h"
 
 #include <Storage/definition.h>
 #include "object_storage.h"
