+++ CompilerSource/parser/parser.cpp
@@ -45,7 +45,7 @@
 using namespace std; //More ease //To interface with externally defined types and functions
 #include "parser_components.h" //duh
 
-#include "general/darray.h"
+#include "darray.h"
 #include "general/parse_basics_old.h"
 #include "object_storage.h"
