+++ ENIGMAsystem/SHELL/Universal_System/Extensions/libpng/Makefile
@@ -1,5 +1,5 @@
 SHARED_SRC_DIR := ../../shared
 override CXXFLAGS += -I$(SHARED_SRC_DIR)/libpng-util
-override LDLIBS += -L$(SHARED_SRC_DIR)/libpng-util -lpng-util -lpng -lz
+override LDLIBS += -lpng -lz
 
-SOURCES += Universal_System/Extensions/libpng/libpng-ext.cpp
+SOURCES += Universal_System/Extensions/libpng/libpng-ext.cpp Universal_System/Extensions/libpng/libpng-util.cpp
