+++ shared/darray.h
