--- shared/event_reader/Makefile
@@ -1,1 +1,0 @@
-SHARED_SOURCES := $(SHARED_SOURCES) $(addprefix event_reader/,$(call relative_wildcard,*.cpp))
