--- shared/eyaml/Makefile
@@ -1,2 +1,0 @@
-# SHARED_INCLUDES += eyaml (Best to just let them include eyaml/eyaml.h)
-SHARED_SOURCES := $(SHARED_SOURCES) $(addprefix eyaml/,$(call relative_wildcard,*.cpp))
