--- shared/rectpacker/Makefile
@@ -1,1 +1,0 @@
-SHARED_SOURCES := $(SHARED_SOURCES) $(addprefix rectpacker/,$(call relative_wildcard,*.cpp))
